pb have to modify

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: jaehejun <jaehejun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/28 22:42:02 by jaehejun          #+#    #+#             */
-/*   Updated: 2023/09/06 15:49:49 by jaehejun         ###   ########.fr       */
+/*   Updated: 2023/09/07 16:45:56 by jaehejun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ static int	is_space(char c)
 	return (0);
 }
 
-int	ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int			sign;
 	long long	number;
@@ -39,5 +39,7 @@ int	ft_atoi(const char *str)
 		number = number * 10 + (*str - '0');
 		str++;
 	}
+	//if (number * sign > 2147483647 || number * sign < -2147483648)
+		
 	return (number * sign);
 }
